piControl: Use pibridge communication interfaces instead of kernel_write/recv
In the linux kernel(github.com/RevolutionPi/linux), an new module pibridge
(compile option:SERIAL_DEV_MUX_PIBRIDGE) has been added. With this commit,
the piControl gets to use this module. In another word, this commit is
depending on the compile option SERIAL_DEV_MUX_PIBRIDGE.

This change is the first step of the refactoring of piControl, in which we
intend to move as many as possible functions to linux kernel according to
the design principles of target subsystems.

Signed-off-by: Lukas Wunner <lukas@wunner.de>
Signed-off-by: z.han <z.han@kunbus.com>